requirements.txt
numpy
pandas
Pillow
opencv-python-headless
streamlit
streamlit-webrtc
tensorflow
